% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsum.R
\name{stat_sum}
\alias{stat_sum}
\title{Summary Statistics}
\usage{
stat_sum(data)
}
\arguments{
\item{data}{is the numeric data in use}

\item{Minimum}{is the minimum of the data}

\item{Maximum}{is the maximum of the data}

\item{Q1}{is the 25\% percentile of the data}

\item{Median}{is the median of the data}

\item{Q3}{is the 75\% percentile of the data}

\item{Range}{is the range of the data}

\item{IQR}{is the interquantile range of the data}

\item{SD}{is the standard deviation of the data}
}
\value{
the numeric minimum, first quartile (Q1), median, third quartile (Q3), maximum, range, and interquartile range (IQR), and standard deviation of the data
}
\description{
Produce summary statistics on the numeric variable/column in the data
}
\examples{
library(gapminder)
Population_Stat <- stat_sum(gapminder$pop)
}
