% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsum.R
\name{stat_summary}
\alias{stat_summary}
\title{Summary Statistics for factors of interest grouped by other factors of interest}
\usage{
stat_summary(data, group_factor, summarize_factor)
}
\arguments{
\item{data}{is the dataset in use}

\item{group_factor}{is the variable/column to be grouped of interest in the data}

\item{summarize_factor}{is the variable/column to be summarized with different summary statistics}
}
\value{
the numeric minimum, first quartile (Q1), median, third quartile (Q3), maximum, range, and interquartile range (IQR), and standard deviation of the factor of interest to be summarized for all groups defined by the grouping factor of interest
}
\description{
Produce summary statistics on the factor of interest to be summarized on each group defined and specified by the grouping factor of interest
}
\examples{
library(datateachr)
neighbourhood_diameter_summary <- stat_summary(vancouver_trees, "neighbourhood_name", "diameter")
}
